I fix the UI bug from the Booking Page, and added a StudentBookingDialog
