+++ src/dialogs/booking/definition.ts
@@ -1,10 +1,18 @@
 import { IBooking } from 'src/entities';
 import { Struct } from 'src/structs';
 
-export type Booking = Struct<
-  'bookAppointmentDialog',
-  {
-    success(booking: IBooking): void;
-    error?: ErrorCallback;
-  }
->;
+export type Booking =
+  | Struct<
+      'bookAppointmentDialog',
+      {
+        success(booking: IBooking): void;
+        error?: ErrorCallback;
+      }
+    >
+  | Struct<
+      'studentBookingDialog',
+      {
+        success: VoidCallback;
+        error?: ErrorCallback;
+      }
+    >;
