+++ src/dialogs/the-dialogs.ts
@@ -5,7 +5,11 @@ import { AccountApplication } from './account-application/definition';
 import { Booking } from './booking/definition';
 import { ApplicationActions } from './admin/definition';
 
-export type DialogStructs = Auth | AccountApplication | ApplicationActions | Booking;
+export type DialogStructs =
+  | Auth
+  | AccountApplication
+  | ApplicationActions
+  | Booking;
 
 type DialogEvents = ToEmit<DialogStructs, DialogStructs>;
 export type WorkflowTypes = ToType<DialogStructs, DialogStructs>;
