+++ src/layouts/MainLayout.vue
@@ -93,6 +93,8 @@
       <ConfirmLockDialog />
       <BookAppointmentDialog />
       <RejectApplicationDialog />
+      <StudentBookingDialog />
+      >
     </q-page-container>
   </q-layout>
 </template>
@@ -121,6 +123,7 @@ import {
 import { theBus } from 'src/the-bus';
 import { date } from 'quasar';
 import LogOutDialog from 'src/dialogs/auth/LogOutDialog.vue';
+import StudentBookingDialog from 'src/dialogs/booking/StudentBookingDialog.vue';
 
 const { locale } = useI18n({ useScope: 'global' });
 const authStore = useAuthStore();
