+++ src/pages/booking/BookingPage.vue
@@ -41,32 +41,21 @@
         <div class="adviser-list-text text-h3 text-bold text-primary q-mb-md">
           Adviser
         </div>
-        <q-table
-          v-if="isBookingAppear"
-          :rows="bookings"
-          :columns="columns"
-          class="q-pa-md q-gutter-sm"
-        >
-          <template #top-right>
-            <q-btn
-              @click="bookAppointment"
-              class="bg-primary text-white"
-              style="margin-right: 10px"
-              >Book</q-btn
-            >
-            <q-btn
-              class="bg-red close-btn text-white"
-              @click="disappearBookingDialog"
-            >
-              CANCEL
-            </q-btn>
-          </template>
-        </q-table>
+
         <div class="adviser-lists row wrap q-gutter-md q-col-gutter-md">
           <q-btn
-            v-for="(adviser, index) in 5"
+            v-for="(adviser, index) in 20"
             :key="index"
-            @click="appearBookingDialog"
+            @click="
+              TheDialogs.emit({
+                type: 'studentBookingDialog',
+                arg: {
+                  success() {
+                    console.log('IM STUDENT');
+                  },
+                },
+              })
+            "
             class="bg-primary q-pa-md"
             :class="['q-md:w-18', 'q-sm:w-45', 'q-xs:w-100']"
           >
@@ -80,51 +69,12 @@
 </template>
 
 <script setup lang="ts">
-import { QTableColumn } from 'quasar';
 import { TheDialogs } from 'src/dialogs/the-dialogs';
-import { IBooking } from 'src/entities';
-import { useBookingStore } from 'src/stores/booking.store';
-import { computed, onMounted, ref } from 'vue';
-const bookingStore = useBookingStore();
-// Define reactive variables
-const isSearchExpanded = ref(false); // Tracks if the search button is expanded
-const isBookingAppear = ref(false); // Tracks if the user is booking an appointment
-const searchQuery = ref<string>(''); // Stores the input value
-const bookings = computed(() => {
-  return bookingStore.bookings;
-});
-const columns = [
-  {
-    name: 'date',
-    label: 'Date',
-    field: 'date',
-  },
-  {
-    name: 'location',
-    label: 'Location',
-    field: 'location',
-  },
-  {
-    author: 'author',
-    label: 'Author',
-    field: (row: IBooking) =>
-      `${typeof row.author == 'object' ? row.author.fullName : ''}`,
-  },
-] as QTableColumn[];
 
-function bookAppointment() {
-  TheDialogs.emit({
-    type: 'bookAppointmentDialog',
-    arg: {
-      success(booking) {
-        console.log(booking);
-      },
-    },
-  });
-}
-onMounted(() => {
-  bookingStore.streamWith();
-});
+import { ref } from 'vue';
+
+const isSearchExpanded = ref(false); // Tracks if the search button is expanded
+const searchQuery = ref<string>(''); // Stores the input value
 
 // Methods
 const expandSearch = () => {
@@ -133,14 +83,6 @@ const expandSearch = () => {
 
 const collapseSearch = () => {
   isSearchExpanded.value = false;
-};
-
-const appearBookingDialog = () => {
-  isBookingAppear.value = true;
-};
-
-const disappearBookingDialog = () => {
-  isBookingAppear.value = false;
 };
 </script>
